requirements.txt
# BAW Project - Wszystkie zależności
# Połączone requirements dla UslugaDoPorownan i SecureDocCompare
# Kompatybilne z Python 3.11-3.13

# ============================================
# Web Framework (wspólne)
# ============================================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# ============================================
# Templates
# ============================================
jinja2>=3.1.4

# ============================================
# HTTP Client
# ============================================
httpx>=0.27.0

# ============================================
# Configuration & Data Validation
# ============================================
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1

# ============================================
# Document Processing (UslugaDoPorownan)
# ============================================
docx2python>=2.0.0
fast-diff-match-patch>=2.0.0

# PDF Conversion (nowa funkcjonalność)
pdf2docx>=0.5.8
pdfplumber>=0.11.0
python-docx>=1.1.0

# ============================================
# Utilities
# ============================================
# Dodatkowe narzędzia jeśli potrzebne
